+++ guillotina/db/storages/cockroach.py
@@ -230,8 +230,8 @@ class CockroachStorage(pg.PostgresqlStorage):
                 'SET DEFAULT_TRANSACTION_ISOLATION TO ' + self._isolation_level)
         return conn
 
-    async def initialize(self, loop=None):
-        await super().initialize(loop=loop)
+    async def initialize(self, loop=None, **kw):
+        await super().initialize(loop=loop, **kw)
         # we need snapshot isolation to allow us to work together with
         # other transactions nicely and prevent deadlocks
         await self._read_conn.execute('SET DEFAULT_TRANSACTION_ISOLATION TO SNAPSHOT')
@@ -300,3 +300,9 @@ class CockroachStorage(pg.PostgresqlStorage):
             logger.warning('Do not have db transaction to commit')
 
         return transaction._tid
+
+    # Cockroach cant use at version 1.0.3 row count (no fetch)
+    async def get_one_row(self, smt, *args):
+        # Helper function to provide easy adaptation to cockroach
+        result = await smt.fetch(*args)
+        return result[0] if len(result) > 0 else None