+++ guillotina/interfaces/__init__.py
@@ -16,8 +16,10 @@ from .content import IContainer  # noqa
 from .content import IContentNegotiation  # noqa
 from .content import IDatabase  # noqa
 from .content import IFolder  # noqa
+from .content import IGetOwner  # noqa
 from .content import IGroupFolder  # noqa
 from .content import IItem  # noqa
+from .content import IJavaScriptApplication  # noqa
 from .content import ILocation  # noqa
 from .content import IRegistry  # noqa
 from .content import IResource  # noqa